Polish Entity-as-Reference framework per review
Substantive: HomeBox attachment proxy returns 502 for upstream
download failures (was 404); EntityEditView logs a warning on
unregistered-integration lookups (was silent); migration 0020 uses
a single bulk delete; CustomTemplateViewData validates a non-empty
template_name on construction.

The canonical HomeBox property -> label table is lifted to a module-
level HB_ITEM_FIELD_PAIRS in shared/hb_converter.py, consumed by both
the Connect resolver and the (dormant) importer. The importer's
synthetic field output broadens from 4 to 10 properties as a result.

Refactors: HbClient._api_url renamed to public api_url; importer-side
tests moved to test_hb_importer.py; removed unreachable rejection
branch in EntityAttributeRegularFormSet.clean(); collapsed nested
attribute gates; stripped a stray template syntax artifact; added
rel='noopener noreferrer' to external links.
