+++ src/hi/apps/entity/forms.py
@@ -47,36 +47,7 @@ class EntityAttributeForm( AttributeForm ):
 
 
 class EntityAttributeRegularFormSet(RegularAttributeBaseFormSet):
-
-    def clean(self):
-        super().clean()
-
-        if not self.instance or self.instance.allow_internal_attributes:
-            return
-
-        for form in self.forms:
-            if not hasattr(form, 'cleaned_data'):
-                continue
-
-            cleaned_data = form.cleaned_data
-            if not cleaned_data:
-                continue
-
-            if cleaned_data.get('DELETE', False):
-                continue
-
-            if form.instance and form.instance.pk:
-                continue
-
-            has_new_data = bool(
-                cleaned_data.get('name')
-                or cleaned_data.get('value')
-                or cleaned_data.get('secret')
-            )
-            if has_new_data:
-                raise forms.ValidationError(
-                    'New attributes cannot be added for this item because attributes are managed externally.'
-                )
+    pass
 
 
 EntityAttributeRegularFormSet = forms.inlineformset_factory(
